add loading component and delete functions
