+++ src/assets/styles/Logs.css
@@ -1,0 +1,7 @@
+#loading-container {
+  display: flex;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+}
