+++ src/pages/Logs/tableColumns/index.ts
@@ -1,0 +1,34 @@
+export const columns = [
+  {
+    label: 'Id',
+    value: 'id',
+  },
+  {
+    label: 'Nome',
+    value: 'name',
+  },
+  {
+    label: 'Classe',
+    value: 'largeGroup',
+  },
+  {
+    label: 'Ordem',
+    value: 'group',
+  },
+  {
+    label: 'Espécie',
+    value: 'species',
+  },
+  {
+    label: 'Tamanho',
+    value: 'length',
+  },
+  {
+    label: 'Massa',
+    value: 'weight',
+  },
+  {
+    label: 'Status',
+    value: 'reviewed',
+  },
+]
